数据库设计完毕，其相关文件已导入项目，但还未修改完善，打算暂存，到我家电脑继续修改，故提交。

+++ zsblogs/src/main/java/com/zs/dao/RoleMapper.java
@@ -1,16 +1,15 @@
 package com.zs.dao;
 
 import com.zs.entity.Role;
-import java.math.BigDecimal;
 
 public interface RoleMapper {
-    int deleteByPrimaryKey(BigDecimal id);
+    int deleteByPrimaryKey(Integer id);
 
     int insert(Role record);
 
     int insertSelective(Role record);
 
-    Role selectByPrimaryKey(BigDecimal id);
+    Role selectByPrimaryKey(Integer id);
 
     int updateByPrimaryKeySelective(Role record);
